Migrate bee allocation system from FlowerPatchController to BiomeRegion
Core Changes:
- Update BeeFleetManager to use BiomeRegion instead of FlowerPatchController for allocation tracking
- Change allocation dictionary key from FlowerPatchController to BiomeRegion
- Update event signatures: UnityEvent<BiomeRegion, int> for OnBeeAllocationChanged
- Rename GetRegisteredFlowerPatchCount() to GetRegisteredRegionCount()

Component Updates:
- RouteController: Remove all FlowerPatchController references, use BiomeRegion exclusively
- BeeController: Cache homePosition from HexTile.BeeGatherPosition to fix return flight height
- HexTile: Set default beeSpawnOffset to Y=2.0 for proper bee height

UI Panel Updates:
- FlowerPatchUpgradePanel: Get BiomeRegion from FlowerPatchController bridge for allocation calls
- FleetManagementPanel: Update OnBeeAllocationChanged signature, filter to show only unlocked regions
- YearStatsTracker: Update method call to GetRegisteredRegionCount()

Architecture:
- FlowerPatchController remains as minimal bridge for UI compatibility
- All core game logic (unlock, capacity, allocation) handled by BiomeRegion
- BeeFleetManager tracks allocations by BiomeRegion directly
- RouteController and BeeController use BiomeRegion for bee spawning

Bug Fixes:
- Fix bee return flight height by caching homePosition from HexTile
- Fix bee spawn height with proper Y offset in HexTile (2.0 instead of 0.5)

+++ Assets/Scripts/Achievement.cs
@@ -48,9 +48,9 @@ public class Achievement : ScriptableObject
     /// <summary>
     /// Get total resources collected across all types.
     /// </summary>
-    private int GetTotalResourcesCount(YearSummary summary)
+    private float GetTotalResourcesCount(YearSummary summary)
     {
-        int total = 0;
+        float total = 0f;
         foreach (var kvp in summary.totalResourcesCollected)
         {
             total += kvp.Value;
